Add ChatAPI about Team

+++ BACKEND/ChatApiServer/src/main/java/com/aomc/coop/service/TeamService.java
@@ -3,6 +3,7 @@ package com.aomc.coop.service;
 import com.aomc.coop.mapper.ChannelMapper;
 import com.aomc.coop.mapper.MessageMapper;
 import com.aomc.coop.mapper.TeamMapper;
+import com.aomc.coop.mapper.UserMapper;
 import com.aomc.coop.model.Channel;
 import com.aomc.coop.model.Message;
 import com.aomc.coop.model.Team;
@@ -35,17 +36,19 @@ public class TeamService {
     @Autowired
     private MessageMapper messageMapper;
 
+    @Autowired
+    private UserMapper userMapper;
+
     private JwtService jwtService;
 
     CodeJsonParser codeJsonParser = CodeJsonParser.getInstance();
+//    ResponseType<T> responseType = new ResponseType<T>();
 
 //    @Resource(name="redisTemplate")
 //    private HashOperations<String, String, String> values;
 
 
     //팀생성
-    //owner를 idx로 넣은것
-    //channel을 배열로 넣는것
     @Transactional
     public ResponseType createTeam(final Team team) {
         try {
@@ -69,28 +72,16 @@ public class TeamService {
             team.setChannels(channels);
 
             teamMapper.createTeam(team);
-
-            System.out.println("createTeam");
-
             channelMapper.createChannel(channel, team.getIdx());
-
-            System.out.println("createChannel " + channel.getIdx());
-
             messageMapper.createMessage(message, channel.getIdx(), users.get(0).getIdx());
-
-            System.out.println("createMessage");
-
 
             for (User user : users) {
                 if(user.getIdx()==team.getOwner()){
                     teamMapper.createUserHasTeam(team.getIdx(), user.getIdx(),1);
-                    System.out.println("createUserHasTeam1");
                 }else{
                     teamMapper.createUserHasTeam(team.getIdx(), user.getIdx(),0);
-                    System.out.println("createUserHasTeam2");
                 }
                 channelMapper.createUserHasChannel(channel.getIdx(), user.getIdx());
-                System.out.println("createUserHasChannel");
             }
 
 
@@ -103,10 +94,32 @@ public class TeamService {
 
     }
 
+    //팀조회
+    public ResponseType readTeam(final int teamIdx) {
+
+            Team team = teamMapper.readTeam(teamIdx);
+
+            if(team == null){
+                return codeJsonParser.codeJsonParser(Status_5000.FAIL_READ.getStatus());
+            }
+            return ResponseType.res(200,"팀 조회 성공", "Success Team Read", team);
+
+    }
+
     //팀수정
     @Transactional
     public ResponseType updateTeam(final Team team) {
         try {
+
+            Team temp = teamMapper.readTeam(team.getIdx());
+
+            if(temp == null){
+                return codeJsonParser.codeJsonParser(Status_5000.FAIL_READ.getStatus());
+            }
+            if(temp.getStatus()==0){
+                return codeJsonParser.codeJsonParser(Status_5000.DEACTIVE_TEAM.getStatus());
+            }
+
             teamMapper.updateTeam(team);
             return codeJsonParser.codeJsonParser(Status_5000.SUCCESS_UPDATE_TEAM.getStatus());
 
@@ -121,14 +134,74 @@ public class TeamService {
     @Transactional
     public ResponseType deleteTeam(final int teamIdx) {
 
-            int delete_flag = teamMapper.deleteTeam(teamIdx);
-            System.out.println(delete_flag);
+        Team temp = teamMapper.readTeam(teamIdx);
 
-            if(delete_flag!=0)
-                return codeJsonParser.codeJsonParser(Status_5000.SUCCESS_DELETE_TEAM.getStatus());
-            else
-                return codeJsonParser.codeJsonParser(Status_5000.FAIL_DELETE_TEAM.getStatus());
+        if(temp == null){
+            return codeJsonParser.codeJsonParser(Status_5000.FAIL_READ.getStatus());
+        }
 
+        if(temp.getStatus()==0){
+            return codeJsonParser.codeJsonParser(Status_5000.DEACTIVE_TEAM.getStatus());
+        }
+        int delete_flag = teamMapper.deleteTeam(teamIdx);
+
+        if(delete_flag!=0)
+            return codeJsonParser.codeJsonParser(Status_5000.SUCCESS_DELETE_TEAM.getStatus());
+        else
+            return codeJsonParser.codeJsonParser(Status_5000.FAIL_DELETE_TEAM.getStatus());
+
+    }
+
+
+    //팀초대
+    @Transactional
+    public ResponseType inviteTeam(final int teamIdx, final String uid) {
+
+
+        User userTemp = userMapper.findBysUserid(uid);
+        if(userTemp ==null){
+            return codeJsonParser.codeJsonParser(Status_5000.FAIL_READ.getStatus());
+        }
+
+        int flag = teamMapper.createUserHasTeam(teamIdx, userTemp.getIdx(), 0);
+
+        if(flag==1)
+            return codeJsonParser.codeJsonParser(Status_5000.SUCCESS_INVITE.getStatus());
+        else
+            return codeJsonParser.codeJsonParser(Status_common.INTERNAL_SERVER_ERROR.getStatus());
+    }
+
+    //채널조회
+    public ResponseType readChannel(final int teamIdx, final int userIdx) {
+
+        List<Channel> channels= channelMapper.readChannel(teamIdx, userIdx);
+
+        if(channels == null){
+            return codeJsonParser.codeJsonParser(Status_5000.FAIL_READ.getStatus());
+        }
+        return ResponseType.res(200,"채널 조회 성공", "Success Channel Read", channels);
+
+    }
+
+    //채널조회
+    @Transactional
+    public ResponseType deactiveUser(final int teamIdx, final int userIdx) {
+
+        int teamFlag = teamMapper.deactiveUserOfTeam(teamIdx, userIdx);
+
+        List<Channel> channels= channelMapper.readChannel(teamIdx, userIdx);
+
+        for (Channel channel : channels) {
+            int channelFlag = channelMapper.deactiveUserOfChannel(channel.getIdx(), userIdx);
+            if(channelFlag!=1){ //예외처리
+                return codeJsonParser.codeJsonParser(Status_5000.FAIL_DEACVITE_USER.getStatus());
+            }
+        }
+
+        if(teamFlag==1)
+            return codeJsonParser.codeJsonParser(Status_5000.SUCCESS_DEACVITE_USER.getStatus());
+        else
+            return codeJsonParser.codeJsonParser(Status_5000.FAIL_DEACVITE_USER.getStatus());
     }
 
 }
